Documentation and refactor
