+++ lib/src/editor/editor.dart
@@ -1,5 +1,5 @@
 export 'providers/canvas_delegate.dart';
 export 'providers/grid_provider.dart';
 export 'providers/zoom_provider.dart';
-export 'ui/tabs.dart';
-export 'ui/wrapper.dart';
+export 'ui/editor.dart';
+export 'ui/editor_tabs.dart';
