+++ lib/src/editor/providers/canvas_delegate.dart
