+++ lib/src/editor/providers/zoom_provider.dart
