+++ lib/src/editor/ui/editor_bottom_bar.dart
