+++ lib/src/navigation/navigation.dart
@@ -1,3 +1,3 @@
 export 'models/organizers.dart';
 export 'ui/navigation_panel.dart';
-export 'ui/tiles.dart';
+export 'ui/navigation_tiles.dart';
