+++ lib/src/utils/utils.dart
@@ -1,3 +1,4 @@
 export 'extensions.dart';
 export 'radii.dart';
+export 'ripple.dart';
 export 'styles.dart';
